Add getData/setData to ensembles and network arrays
These are for attaching generic information to nodes.  For example,
affiliating an hrr.Vocabulary for display purposes, or an x/y/z
location, or information needed for the BG/Thalamus displays.  This
will reduce the reliance on global variables (and the associated
memory leaks) and eliminate the weird abuse of the documentation
string.